third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/interfaces/TextTrack/activeCues-expected.txt

This is a testharness.js-based test.
FAIL TextTrack.activeCues, empty list assert_not_equals: t1.activeCues and t2.activeCues should be different objects got disallowed value null
FAIL TextTrack.activeCues, after addCue() assert_equals: t1.activeCues should return same object expected null but got object "[object TextTrackCueList]"
FAIL TextTrack.activeCues, different modes assert_equals: t1.activeCues should return the same object after setting mode to showing expected null but got object "[object TextTrackCueList]"
FAIL TextTrack.activeCues, video loading assert_equals: t1.activeCues should return the same object after loading a video expected null but got object "[object TextTrackCueList]"
FAIL TextTrack.activeCues, video playing assert_unreached: Reached unreachable code
FAIL TextTrack.activeCues, adding cue during playback assert_unreached: Reached unreachable code
Harness: the test ran to completion.
